pyproject.toml
[project]
name = "sh2mp4"
description = "Records shell commands to MP4 videos"
version = "0.3.5"
authors = [
    { name = "Gareth Davidson", email = "gaz@bitplane.net" }
]
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    "Pillow>=9.0.0",
    "pytk"
]

[project.optional-dependencies]
dev = [
    "pre-commit",
    "pytest",
    "coverage",
    "pytest-cov",
    "pytest-asyncio",
    "build",
    "twine",
    "ruff",
    "pydoc-markdown"
]

[project.scripts]
sh2mp4 = "sh2mp4.__main__:main"

[build-system]
build-backend = "flit_core.buildapi"
requires = ["flit_core >=3.2,<4"]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.format]
docstring-code-format = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
